Debug: Move debug box code to separate file
